[MOV] mrp, (purchase_)stock: move _get_route_domain to mixin
