+++ addons/mrp/models/stock_replenish_mixin.py
@@ -16,11 +16,6 @@ class ProductReplenishMixin(models.AbstractModel):
             res['bom_id'] = self.env['mrp.bom']._bom_find(product_id)[product_id].id
         return res
 
-    # @api.depends('product_id')
-    # def _compute_bom(self):
-    #     for rec in self:
-    #         rec.bom_id = self.env['mrp.bom']._bom_find(rec.product_id)[rec.product_id]
-
     @api.depends('route_id')
     def _compute_show_bom(self):
         for rec in self:
